Remove deletion of all task delegates in URLSessionDidFinishEventsForBackgroundURLSession:
This should not happen because sometimes this method is called after waking the app up to
respond to authentication challenges and no tasks have completed yet.  It is not safe to
assume that all tasks are complete and the delegates should be removed.  Doing some
results in orphaned tasks with AF delegates to handle them.

Task delegates already get deleted after the URLSession:task:didCompleteWithError: call
which will be made for each task.

Issue: 1579